+++ eagleboard-api/src/main/java/com/mass3d/metadata/version/MetadataVersionService.java
@@ -36,7 +36,7 @@ public interface MetadataVersionService {
    * @param id the identifier.
    * @param name the name.
    */
-  void updateVersionName(int id, String name);
+  void updateVersionName(long id, String name);
 
   /**
    * @param version Version object to delete.
@@ -49,7 +49,7 @@ public interface MetadataVersionService {
    * @param id Key to lookup the value with.
    * @return Version that matched key, or null if there was no match.
    */
-  MetadataVersion getVersionById(int id);
+  MetadataVersion getVersionById(long id);
 
   /**
    * @return List of all version objects.
