+++ eagleboard-api/src/main/java/com/mass3d/metadata/version/MetadataVersionStore.java
@@ -17,7 +17,7 @@ public interface MetadataVersionStore
    * @param id Key to lookup.
    * @return MetadataVersion Value that matched key, or null if there was no match.
    */
-  MetadataVersion getVersionByKey(int id);
+  MetadataVersion getVersionByKey(Long id);
 
   /**
    * Get the version by name.
